refactor: modularize function
